Refactored the client class.
